requirements.txt
feedparser
requests
beautifulsoup4
google-generativeai
newspaper3k
lxml_html_clean